+++ cjs/lib/__bundle-895d88d5.cjs
@@ -1,1 +1,0 @@
-var e=require("./__bundle-97302804.cjs");class s extends e.BaseCommand{}exports.NotificationCollectionRefreshCommand=s,exports.getFeedChannelIndexBy=e=>["-lastMessageUpdatedAt","-createdAt","syncIndex"];