+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## v4.12.1 (Apr 11, 2024)
+### **Features**
+- Added additional parameters in `GroupChannelEventContext`
+
+### **Improvement**
+- Fixed a bug that `lastMessage` not updating on reply
+- Fixed a bug that database upgrade fails in certain environment
+- Improvement stability
+## vv4.12.0 (Mar 25, 2024)
+### **Improvement**
+- Added `markPushNotificationAsDelivered`
+- Added `markPushNotificationAsClicked`
+- Added token registration with device info
+- Added `logViewed`/`logClicked` in FeedChannel
+- Deprecated `markAsClicked`/`logImpression` in FeedChannel
+- (internal)Refactoring statCollector
+- Fixed a bug that `onMessagesUpdated` event not firing on ThreadInfo updated event
+- Exported `SendbirdErrorCode`
+- Improvement stability
 ## v4.11.3 (Mar 14, 2024)
 ### **Improvement**
 - Added `keys` property to `MessageTemplateListParams`